requirements/base.txt
fastapi>=0.109.2
pydantic-settings==2.0.0
pypdf==4.2.0
python-multipart==0.0.10
tortoise-orm>=0.16.21
aerich>=0.5.8
asyncpg>=0.24.0
dependency_injector>=4.0
